LowpanService: Reconstitute mIpManager every time we attach
I noticed that we only seem to be making it past the ObtainingIpState
if it is the first time we have entered the state. After successfully
getting past this state once, if we bring the interface down and then
bring it back up we will be stuck in that state (and, thus, have no
connectivity).

Forcing IpManager to be reconstituted with every attempt fixes the
problem.

Bug: b/67718802
Test: manual
Change-Id: Idb06068c721c2e432c9936d6a00b21ab8ea6febc